requirements.txt
streamlit==1.33.0
pandas==2.0.3
SQLAlchemy==2.0.29
python-dotenv==1.0.1